Drag&DropComponents

+++ api/Domain/ProjectColumn.cs
@@ -7,7 +7,7 @@ namespace Domain
     public class ProjectColumn
     {
         public int Id { get; set; }
-        public int Order { get; set; }
+        public double Order { get; set; }
         public string Title { get; set; }
         public int ProjectId { get; set; }  
         public Project Project { get; set; }
